Import first version of uds_rw
The first version is partially tested, with only a diagnostic test,
for an ECU update, read DID, write DID, and DTC handling.

It allows encoding and decoding of UDS messages, not the whole set,
but a big part of it.

It can be correlated with Rust crates doip_rw and doip_rw_tokio to
have a full diagnostic stack over IPv4.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>